// frontend/src/app/api/model/employeeUpdateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeUpdateDto { 
    firstName: string;
    lastName: string;
    email: string;
    salary: number;
    officeId?: number;
}