// frontend/src/app/api/model/employeeViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeViewDto { 
    id?: string;
    username?: string;
    email?: string;
    firstName?: string;
    lastName?: string;
    employeeNumber?: string;
    hireDate?: string;
    salary?: number;
    applicationRole?: EmployeeViewDto.ApplicationRoleEnum;
    isActive?: boolean;
    isEmailVerified?: boolean;
    officeId?: number;
}
export namespace EmployeeViewDto {
    export const ApplicationRoleEnum = {
        Client: 'CLIENT',
        Courier: 'COURIER',
        Clerk: 'CLERK',
        Admin: 'ADMIN'
    } as const;
    export type ApplicationRoleEnum = typeof ApplicationRoleEnum[keyof typeof ApplicationRoleEnum];
}